refactor(backend): rename service-configuration module without .schema suffix
Rename service-configuration.schema.ts to service-configuration.ts. Update imports,
.env.example, and 004 spec task paths.

Enable VS Code explorer file nesting for *.ts next to matching *.spec.ts files.

Made-with: Cursor
