+++ apps/backend/.env.example
@@ -1,4 +1,4 @@
-# Service configuration (see src/config/service-configuration.schema.ts).
+# Service configuration (see src/config/service-configuration.ts).
 # Copy to .env and load into the environment, or export vars in your shell.
 # This app does not load .env files by default; Nest reads process.env only.
